+++ glide/estimators/__init__.py
@@ -2,6 +2,7 @@ from glide.estimators.asi import ASIMeanEstimator
 from glide.estimators.classical import ClassicalMeanEstimator
 from glide.estimators.ipw_classical import IPWClassicalMeanEstimator
 from glide.estimators.ppi import PPIMeanEstimator
+from glide.estimators.ptd import PTDMeanEstimator
 from glide.estimators.stratified_classical import StratifiedClassicalMeanEstimator
 from glide.estimators.stratified_ppi import StratifiedPPIMeanEstimator
 
@@ -10,6 +11,7 @@ __all__ = [
     "IPWClassicalMeanEstimator",
     "StratifiedClassicalMeanEstimator",
     "PPIMeanEstimator",
+    "PTDMeanEstimator",
     "ASIMeanEstimator",
     "StratifiedPPIMeanEstimator",
 ]