+++ tests/unit/estimators/test_ppi.py
@@ -2,6 +2,7 @@ import numpy as np
 import pytest
 from numpy.typing import NDArray
 
+from glide.confidence_intervals import CLTConfidenceInterval
 from glide.core.mean_inference_result import SemiSupervisedMeanInferenceResult
 from glide.estimators.ppi import PPIMeanEstimator
 
@@ -111,6 +112,7 @@ def test_estimate_is_valid_inference_result(estimator, y_arrays):
     y_true, y_proxy = y_arrays
     result = estimator.estimate(y_true, y_proxy)
     assert isinstance(result, SemiSupervisedMeanInferenceResult)
+    assert isinstance(result.confidence_interval, CLTConfidenceInterval)
     assert np.isfinite(result.confidence_interval.lower_bound)
     assert np.isfinite(result.confidence_interval.upper_bound)
     assert result.confidence_interval.lower_bound < result.confidence_interval.upper_bound